reduce change, make private functions private
